requirements.txt
PyQt5==5.14.0
qdarkstyle
python-rapidjson
